requirements.txt
Django
djangorestframework
django-cors-headers 
Pillow == 7.0
django-filter